penambahan login.php & logout.php, dan harga.php

+++ laundry/delete_bulan.php
@@ -1,5 +1,10 @@
 <?php
+session_start();
 include 'db.php';
+if (!isset($_SESSION['username'])) {
+  header("Location: login.php");
+  exit();
+}
 
 $delete = mysqli_query($conn, "DELETE FROM data_bulan WHERE id = '".$_GET['id']."'");
